requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.2
factory_boy==3.3.1
gunicorn==22.0.0
packaging==24.0
python-dotenv==1.0.1
responses==0.25.3
sqlparse==0.5.0
typing_extensions==4.11.0
pre-commit==3.5.0